Update 20260503 - Rename Arrowhead to Bolt Head and UI Fixes
Replace Arrowhead with BoltHead trap and add related assets, data and registration changes. Renamed Arrowhead classes/resources to BoltHead (block, item, properties, damage type, lang entries, models and blockstates), added bolt_head assets and recipes, and updated ModBlocks/ModItems/ModItemGroups accordingly. Fix RegistryHelper.registerStat to use the correct registry key ordering. Improve recipe/crafting UX: enhance CustomShapedRecipe to match recipes at offsets (allow centering in larger grids), update BaseCraftingScreenHandler to center and fill input slots for centered recipes, and add centered ghost-recipe rendering in BlueprintWidget. Adjust BaseRecipeBookWidget to use dynamic UV/texture dimensions, handle ModRecipeBookType categories, remove unnecessary null checks, and tweak input/search handling. Minor fixes: casting in BaseAnimatedResultButton sin call, small API/cleanup changes (TurretAssemblyStationBlock spacing, ModStats additions, TODO update). These changes enable bolt-head traps and improve recipe alignment/recipe-book behavior for the Turret Assembly Station UI and custom crafting logic.

This is commit finally implement TAS (#35) and all its intended features.

[CHANGELOG]
🟢 Added access widener file alongside an access widener entry.
🟢 Created several UI super classes for implementation of the custom "Recipe Book" UI (called Blueprint).
🟢 Added several new textures for the custom UI.
🟢 Added Turret Assembly Station recipe and recipe advancement.
🟢 Added new turret tag.
🟡 Updated several UI classes to now use the newly created super classes.
🟡 Refactored several recipe paths after grouping each crafting recipe into their respective group.
🟡 Fixed unaligned ghost recipes.
🟡 Fixed recipe advancements triggering on wrong conditions.
🟡 Adjusted Blueprint widget's left offset.
🟡 Updated TODO file.

+++ src/client/java/com/virus5600/defensive_measures/gui/screen/book/BaseAnimatedResultButton.java
@@ -91,7 +91,7 @@ public class BaseAnimatedResultButton extends ClickableWidget {
 		boolean shouldBounce = this.bounce > 0.0F;
 
 		if (shouldBounce) {
-			float f = 1.0F + 0.1F * (float)Math.sin((double)(this.bounce / 15.0F * (float)Math.PI));
+			float f = 1.0F + 0.1F * (float)Math.sin((this.bounce / 15.0F * (float)Math.PI));
 			ctx.getMatrices().pushMatrix();
 			ctx.getMatrices().translate((float)(this.getX() + 8), (float)(this.getY() + 12));
 			ctx.getMatrices().scale(f, f);
